made the form into a component
made the form into a component and added a modal for PE
