+++ src/routes/addPartner/+page.svelte
@@ -1,17 +1,7 @@
 <script>
+	import PartnerForm from "../../lib/components/partnerForm.svelte";
+
 	export let form;
-	
-	function submitted() {
-		if (form?.success) {
-            alert(form?.message);
-            setTimeout( () => {
-                window.location.href = '/'
-            }, 1000)
-			
-		}else if(form?.success == false){
-            alert(form?.message);
-        }
-	}
 </script>
 
 <section class="content-container">
@@ -20,16 +10,7 @@
 		<p>Probeer een partner toe te voegen die nog niet in de lijst bestaat!</p>
 	</article>
 
-	<section class="form-container">
-		<form on:submit={submitted()} method="POST">
-			<label for="name">Voeg een partner titel toe</label>
-			<input id="name" name="name" required type="text" />
-
-			<label for="url" class="url-label">Voeg een Partner homepagina URL toe</label>
-			<input id="url" name="url" required type="url" />
-			<button class="add-button">Partner toevoegen</button>
-		</form>
-	</section>
+	<PartnerForm {form}/>
 </section>
 
 <style>
@@ -40,8 +21,7 @@
 		align-items: flex-start;
 	}
 
-	article,
-	.form-container {
+	article {
 		background-color: var(--c-container);
 		margin: 0.5em;
 		border-radius: 4px;
@@ -52,53 +32,9 @@
 		padding: 1.2em;
 	}
 
-	.form-container {
-		line-height: 2em;
-		width: 100%;
-	}
-
 	.tip {
 		color: var(--c-orange);
 		line-height: 1.5em;
-	}
-
-	.add-button {
-		border: none;
-		background-color: var(--c-modal-button);
-		color: white;
-		padding: 0.7em;
-		cursor: pointer;
-		text-decoration: none;
-		transition: 0.3s;
-		border-radius: 4px;
-		margin-top: 2em;
-	}
-
-	input[type='text'],
-	input[type='url'] {
-		width: 100%;
-		padding: 12px 20px;
-		display: inline-block;
-		border: 1px solid #ccc;
-		border-radius: 4px;
-		box-sizing: border-box;
-		max-width: 700px;
-	}
-
-	.add-button:hover {
-		opacity: 0.75;
-	}
-
-	.url-label {
-		margin-top: 1em;
-	}
-
-	form {
-		width: 100%;
-		display: flex;
-		flex-direction: column;
-		justify-content: flex-start;
-		align-items: flex-start;
 	}
 
 	@media (max-width: 850px) {
